scsi: lpfc: Ignore GID-FT response that may be received after
From 04c1d9c50ae32d6efd0b71024b3829051821c7a2 Mon Sep 17 00:00:00 2001
 a link flip
Git-commit: 04c1d9c50ae32d6efd0b71024b3829051821c7a2
Patch-mainline: v5.14-rc1
References: bsn#61

When a link bounce happens, there is a possibility that responses to
requests posted prior to the link bounce could be received. This is
problematic as the counter to track reglogin completion after link up can
become out of sync with the real state.

As there is no reason to process a request made in a prior link up context,
eliminate all the disturbance by tagging the request with the event_tag
maintained by the SLI Port for the link. The event_tag will change on every
link state transition.  As long as the tag matches the current event_tag,
the response can be processed. If it doesn't match, just discard the
response.

Link: https://lore.kernel.org/r/20210514195559.119853-8-jsmart2021@gmail.com
Co-developed-by: Justin Tee <justin.tee@broadcom.com>
Signed-off-by: Justin Tee <justin.tee@broadcom.com>
Signed-off-by: James Smart <jsmart2021@gmail.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.suse/v5.14-rc1-scsi-lpfc-Ignore-GID-FT-response-that-may-be-receive.patch