mlxsw: Verify the accessed index doesn't exceed the array
Git-commit: 33d89128a9602c34d2875dbf7153d9bdbccd7511
 length
References: bsn#19
Patch-mainline: v5.10.83

[ Upstream commit 837ec05cfea08284c575e8e834777b107da5ff9d ]

There are few cases in which an array index queried from a fw register,
is accessed without any validation that it doesn't exceed the array
length.

Add a proper length validation, so accessing memory past the end of an
array will be forbidden.

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.83-0076-mlxsw-Verify-the-accessed-index-doesn-t-exceed-the-a.patch