ARM: dts: OMAP2+: Replace underscores in sub-mailbox node
Git-commit: 28d2ae98154e470ab9bd0e36d556d02e8119b1ad
 names
References: bsn#19
Patch-mainline: v5.10.53

[ Upstream commit 9e7f5ee1137397def6580461e27e5efcb68183ee ]

A number of sub-mailbox node names in various OMAP2+ dts files are
currently using underscores. This is not adhering to the node name
convention, fix all of these to use hiphens.

These nodes are already using the prefix mbox, so they will be in
compliance with the sub-mailbox node name convention being added in
the OMAP Mailbox YAML binding as well.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.53-0032-ARM-dts-OMAP2-Replace-underscores-in-sub-mailbox-nod.patch